requirements.txt
Django==1.7
argparse==1.2.1
djangorestframework==2.4.3
djangorestframework-gis==0.7
gpsbabel==1.1.1
numpy==1.9.0
psycopg2==2.5.4
wsgiref==0.1.2
